use int instead of float to have less type conversions

+++ model_objective_stop_balancing.go
@@ -14,15 +14,15 @@ func (t *balanceObjectiveImpl) EstimateDeltaValue(
 	solution := move.Solution()
 	oldMax := t.Max(solution, nil)
 	NewMax := t.Max(solution, move)
-	return (NewMax - oldMax)
+	return float64(NewMax - oldMax)
 }
 
 func (t *balanceObjectiveImpl) Value(solution Solution) float64 {
-	return t.Max(solution, nil)
+	return float64(t.Max(solution, nil))
 }
 
-func (t *balanceObjectiveImpl) Max(solution Solution, move SolutionMoveStops) float64 {
-	max := 0.0
+func (t *balanceObjectiveImpl) Max(solution Solution, move SolutionMoveStops) int {
+	max := 0
 	moveExists := move != nil
 	var vehicle SolutionVehicle
 	if moveExists {
@@ -30,12 +30,12 @@ func (t *balanceObjectiveImpl) Max(solution Solution, move SolutionMoveStops) fl
 	}
 
 	for _, v := range solution.Vehicles() {
-		if max < float64(v.NumberOfStops()) {
-			max = float64(v.NumberOfStops())
+		if max < v.NumberOfStops() {
+			max = v.NumberOfStops()
 		}
 		if moveExists && v.Index() == vehicle.Index() {
-			if max < float64(v.NumberOfStops()+move.StopPositionsLength()) {
-				max = float64(v.NumberOfStops() + move.StopPositionsLength())
+			if max < v.NumberOfStops()+move.StopPositionsLength() {
+				max = v.NumberOfStops() + move.StopPositionsLength()
 			}
 		}
 	}
